move some common code out
