+++ pkg/test/v3/server.go
@@ -17,11 +17,6 @@
 package test
 
 import (
-	"context"
-	"fmt"
-	"log"
-	"net/http"
-
 	"google.golang.org/grpc"
 
 	accessloggrpc "github.com/envoyproxy/go-control-plane/envoy/service/accesslog/v3"
@@ -34,31 +29,6 @@ import (
 	secretservice "github.com/envoyproxy/go-control-plane/envoy/service/secret/v3"
 	"github.com/envoyproxy/go-control-plane/pkg/server/v3"
 )
-
-const (
-	// Hello is the echo message
-	Hello = "Hi, there!\n"
-)
-
-type echo struct{}
-
-func (h echo) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/text")
-	if _, err := w.Write([]byte(Hello)); err != nil {
-		log.Println(err)
-	}
-}
-
-// RunHTTP opens a simple listener on the port.
-func RunHTTP(ctx context.Context, upstreamPort uint) {
-	log.Printf("upstream listening HTTP/1.1 on %d\n", upstreamPort)
-	server := &http.Server{Addr: fmt.Sprintf(":%d", upstreamPort), Handler: echo{}}
-	go func() {
-		if err := server.ListenAndServe(); err != nil {
-			log.Println(err)
-		}
-	}()
-}
 
 // RegisterAccessLogServer starts an accessloggrpc service.
 func RegisterAccessLogServer(grpcServer *grpc.Server, als *AccessLogService) {
